Replace pycrypto with pycryptodome
This replaces `pycrypto` requirement with `pycryptodome`. It is a
drop-in replacement for `pycrypto` which is not actively maintained and
has a known vulnerability, `CVE-2013-7459`. Moreover, one of the
requirements of this package, `pdfminer.six` depends on `pycryptodome`,
and `pycryptodome` conflicts with `pycrypto`, so these packages will
interfere when they are installed simultaneously. All examples have been
tested and run without errors